Replace {{ headnode_public_ip }} with '*' for vhost directive
This type of configuration is used in almost all name-based
virtual hosting situations.

You may replace * with a specific IP address but such virtual
hosts will only be used for HTTP requests received on connection
to the specified IP address. Additionally it is useful to use *
on systems where the IP address is not predictable.