The shared_jobs option allows multiple jobs to share compute nodes.
This information is used by the Job Performance Data (SUPReMM) module to determine which
HPC jobs shared compute nodes. The default is that resources are assumed to not allow node
sharing. If the SUPReMM module is in use and a resource does allow node sharing then this
should be set to true.